requirements.txt
streamlit
langgraph
rdflib
faiss-cpu
sentence-transformers
numpy
pytesseract
pdf2image
Pillow
requests
pandas
